fix call bug in rega
When rdx is used to return a value and is used as argument,
it is in the call defs and hence made dead by the loop
modified here.  This is obviously erroneous behavior.
We instead rephrase the loop to make it clear that among
the caller-save registers, only the ones used by the call
must be live before the call.