simplify a buggy test
I found it by compiling -O2 and seeing the ABI code
fail.  Further investigation revealed GCC trimmed
away the last iteration of the loop because I was
accessing the third element of an array of size two.
This is undefined behavior, so GCC "proved" that the
last iteration was never run.