accept "ret" for functions with a return type
This happens to be needed for C.  The standard mandates that a return
value is used if the caller uses it.  Surprisingly, if the return
"value" is not used, the callee can use "return;".

A better solution is to add an "undef" value and return it, "undef"
would also have other use cases for compiling C.