loosen assertion in load elimination
The assertion was invalid, I was assuming il->blk was
b when writing it.  The new assertion should be right:
If the loop level were to decrease we would get out of
a cycle in cfg, this should not be possible unless we
go through a block with more than 1 predecessor.