improve the range of action of load elimination
When eliminating `load %foo`, don't limit the
search to the live range of %foo, but to the
live range of its aliasing information.

For example, if %foo is a constant offset into
a stack-allocated slot,

    %foo =l add %slot, 42

the search will proceed on all the code in
which %slot is live, not only below the
definition of %foo, like before.