enable arm64 tests
The vararg tests had to be changed because
va_list is 32-bit wide on arm.  The astute
reader will notice that the way we pass
va_list values is wrong, we should be using
the ':valist' type as defined below instead
of 'l'.  But eh, that works for now, because
of the ABI.

    type :valist = align 8 { 32 }