2 bug fixes in rega
The worst one was that "part 3" of rega()
could break the critical invariant that
two interferring temporaries get assigned
different registers.  This is fixed by
being careful when changing the register
of a temporary based on predecessor
blocks.

Thanks to Michael Forney for reporting
these bugs and helping with the analysis.