Let runtime crash on zero div, don't fold it.
Remarks from Quentin:

It is an important decision to use Bot
and not Top as the result of 'x / 0'.
By using Bot, we refuse to give a warrant
to the compiler that would allow meaningless
subsequent decisions.  An example follows.

Clang, on my computer, will build a program
which prints "Ho" when fed the following C:

    int main()
    { puts(1/0 ? "Hi" : "Ho"); }

On the other hand, a C compiler based on
QBE will build a program which crashes,
as one would expect.

See also
https://c9x.me/notes/2014-09-10.html